+++ tests/acceptance/settings/profile/names-test.ts
@@ -12,7 +12,6 @@ module('Acceptance | settings | profile | name', hooks => {
 
     test('cannot use unauthenticated', async function(assert) {
         this.owner.register('service:current-user', CurrentUserStub);
-        server.create('root', { currentUser: null });
         const currentUser = this.owner.lookup('service:current-user');
         await visit('settings/profile/name');
         assert.equal(currentUser.urlCalled, '/settings/profile/name');