+++ types/ember-cp-validations/index.d.ts
@@ -48,4 +48,4 @@ interface Validations {
 }
 
 export function buildValidations(validations: any, globalOptions?: any): Validations;
-export function validator(attrName: string, message: any): any;
+export function validator(attrName: string, options?: any): any;